Consider 'tags' configuration when counting tags
Previously all tags present for all relevant pages are counted,
regardless of the plugin configuration. For example, having a blog post
with the tags 'coffee' and 'tea' and limiting the tags in the
configuration:

    tags = ["coffee"]

The tag counts would still contain the tag 'tea'. This patch counts only
tags that are configured for the plugin.