Fix test breakage due to changes in Lektor admin server code
tests/test_lektor_tag.py::test_resolver was using
non-public/non-stable API for the Lektor admin server.  (That API has
changed enough in the 3.3 branch to minorly break the test.  The API
changes drastically in the master/3.4 branch of lektor.)

Here we refactor the test to avoid use of the Lektor admin server.