Improve tag counting (#25)
* Make counting tags robust against duplicates

If a page has duplicate tags, the tag count previously was increased for
each duplicate. Since this could skew the result, ignore duplicates when
counting tags.

* Consider 'tags' configuration when counting tags

Previously all tags present for all relevant pages are counted,
regardless of the plugin configuration. For example, having a blog post
with the tags 'coffee' and 'tea' and limiting the tags in the
configuration:

    tags = ["coffee"]

The tag counts would still contain the tag 'tea'. This patch counts only
tags that are configured for the plugin.